Server erhält nachrichten von Client auch wenn sie seltsam aussehen

+++ Client/vehicleclient.h
@@ -8,6 +8,7 @@
 #include <iostream>
 #include <getopt.h>
 #include <unistd.h>
+#include <errno.h>
 
 #define KEY 14049
 #define KEY2 14062	/* eindeutiger Key z.B. Matrikelnummer */
@@ -17,6 +18,6 @@
 
 typedef struct
 {
-	long mType;
+	int mType;
 	int mText;
 } message_t;
